refactor: rename model module and update imports; add error handling and helper classes
