+++ pyproject.toml
@@ -25,7 +25,7 @@ Repository = "https://github.com/RimuEirnarn/sqlite_database.git"
 
 [tool.setuptools]
 zip-safe = true
-packages = ["sqlite_database", "sqlite_database.model"]
+packages = ["sqlite_database", "sqlite_database.models"]
 
 [project.optional-dependencies]
 dev = ["pytest", "pylint", "black"]
